refactor: rename `Dashboard` components folder into `Me`
