+++ src/pages/me/proposals/[id].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
-import { ProposalForm } from "@/components/Dashboard/ProposalForm";
-import { Layout } from "@/components/Dashboard/Layout";
+import { ProposalForm } from "@/components/Me/ProposalForm";
+import { Layout } from "@/components/Me/Layout";
 import { api } from "@/utils/api";
 
 const ProposalPage = () => {
